Loading saves finally works
Yay, loading a save should now fully work. Taking into account both card order, but also pressure.
While I was at it I also did some tweaks in the save file to store the current seed of the game. While with the current save structure it doesn't really matter, it will affect the rest of the game later down the line and perhaps I won't have to save the order of every card.